Add mutex in one-thread list processing

+++ include/num_type.h
@@ -1,6 +1,6 @@
 #ifndef NUM_TYPE_H
 #define NUM_TYPE_H
 
-typedef unsigned NUM_TYPE;
+typedef short unsigned NUM_TYPE;
 
 #endif // NUM_TYPE_H
